clean up tests files for now
